abstract common functions away in data handlers
