+++ asrtoolkit/data_handlers/data_handlers_common.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+"""
+Module for common utils for file/data handlers
+"""
+
+separator = "\n"
+
+
+def header():
+  " Returns header "
+  return ""
+
+
+def footer():
+  " Returns footer "
+  return ""
