+++ asrtoolkit/data_handlers/srt.py
@@ -5,23 +5,12 @@ Module for reading/writing SRT files
 This expects a segment from class derived in convert_text
 """
 
-import re
 from webvtt import WebVTT
-from asrtoolkit.data_structures.segment import segment, seconds_to_timestamp
+from asrtoolkit.data_handlers.webvtt_common import read_caption
+from asrtoolkit.data_structures.segment import seconds_to_timestamp
 
-non_transcript_marks = re.compile(r"\[[A-Za-z0-9]{1,}\]")
-
-separator = "\n"
-
-
-def header():
-  " Returns header "
-  return ""
-
-
-def footer():
-  " Returns footer "
-  return ""
+# do not delete - needed in time_aligned_text
+from asrtoolkit.data_handlers.data_handlers_common import separator, header, footer
 
 
 def format_segment(seg):
@@ -34,26 +23,6 @@ def format_segment(seg):
   ret_str += "{:}\n\n".format(seg.formatted_text if seg.formatted_text else seg.text)
 
   return ret_str
-
-
-def read_caption(caption):
-  """
-    Parses caption object to return a segment object
-  """
-  seg = None
-
-  try:
-    start = caption.start_in_seconds
-    stop = caption.end_in_seconds
-
-    text = re.sub(non_transcript_marks, lambda v: "", caption.text.strip()).strip()
-
-    seg = segment({'start': start, 'stop': stop, 'text': text})
-  except Exception as exc:
-    seg = None
-    print(exc)
-
-  return seg if seg and seg.validate() else None
 
 
 def read_file(file_name):
